Separate out test classes
Though they both have to connecting, the tested portions are wholly separate and thus do not need to be kept together
